+++ packages/forms/src/validators/isUrl.d.ts
@@ -1,1 +1,3 @@
-export default function (errorMessage?: string | ((...arg) => string)): string | undefined;
+export default function (
+    errorMessage?: string | ((...arg) => string),
+): (...arg) => string | undefined;