+++ packages/forms/src/validators/max.d.ts
@@ -1,4 +1,4 @@
 export default function (
     maxValue: number,
     errorMessage?: string | ((...arg) => string),
-): string | undefined;
+): (...arg) => string | undefined;