+++ packages/listview/src/components/ColumnHeaderFilterText/FilterOverlay/manageTab/index.d.ts
@@ -1,0 +1,2 @@
+declare function manageTab(node: HTMLElement, event: React.KeyboardEvent<Element>): void;
+export default manageTab;