+++ packages/app/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.55.0](https://github.com/nexxtway/rainbow-modules/compare/v0.54.0...v0.55.0) (2022-11-03)
+
+### Bug Fixes
+
+-   rainbow imports and resolvers ([#636](https://github.com/nexxtway/rainbow-modules/issues/636)) ([e4080d8](https://github.com/nexxtway/rainbow-modules/commit/e4080d8a8f79de8352e3ac944a8a7e1b274912e8))
+
 # [0.54.0](https://github.com/nexxtway/rainbow-modules/compare/v0.53.0...v0.54.0) (2022-10-19)
 
 ### Bug Fixes